# public/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "City Shop Directory API",
        "description": "API documentation for the City Shop Directory mobile application",
        "contact": {
            "name": "API Support",
            "email": "admin@cityshops.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/v1/cities": {
            "get": {
                "tags": [
                    "Cities"
                ],
                "summary": "Get all cities",
                "operationId": "10228b26f3f65f7349e87125b7a96dfa",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search by city name, country, or state",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "country",
                        "in": "query",
                        "description": "Filter by country",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "state",
                        "in": "query",
                        "description": "Filter by state",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number for pagination",
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Cities list",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "data": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/City"
                                                    }
                                                },
                                                "current_page": {
                                                    "type": "integer"
                                                },
                                                "last_page": {
                                                    "type": "integer"
                                                },
                                                "per_page": {
                                                    "type": "integer"
                                                },
                                                "total": {
                                                    "type": "integer"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/cities/{city}": {
            "get": {
                "tags": [
                    "Cities"
                ],
                "summary": "Get single city with shops",
                "operationId": "a30304a1263217b7b28eaa0d159dc193",
                "parameters": [
                    {
                        "name": "city",
                        "in": "path",
                        "description": "City ID or slug",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "City details with shops",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/City"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "City not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "City not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/cities/{city}/featured-shops": {
            "get": {
                "tags": [
                    "Cities"
                ],
                "summary": "Get featured shops for a city",
                "description": "Returns paginated list of featured shops for a specific city",
                "operationId": "5121dc59091b58adfd3ff8b7d34011ee",
                "parameters": [
                    {
                        "name": "city",
                        "in": "path",
                        "description": "City ID or slug",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Number of shops per page",
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Featured shops list",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "shops": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer"
                                                            },
                                                            "name": {
                                                                "type": "string"
                                                            },
                                                            "slug": {
                                                                "type": "string"
                                                            },
                                                            "description": {
                                                                "type": "string"
                                                            },
                                                            "rating": {
                                                                "type": "number",
                                                                "format": "float"
                                                            },
                                                            "review_count": {
                                                                "type": "integer"
                                                            },
                                                            "featured_until": {
                                                                "type": "string",
                                                                "format": "date-time"
                                                            },
                                                            "category": {
                                                                "properties": {
                                                                    "id": {
                                                                        "type": "integer"
                                                                    },
                                                                    "name": {
                                                                        "type": "string"
                                                                    },
                                                                    "icon": {
                                                                        "type": "string"
                                                                    }
                                                                },
                                                                "type": "object"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "total": {
                                                    "type": "integer"
                                                },
                                                "current_page": {
                                                    "type": "integer"
                                                },
                                                "last_page": {
                                                    "type": "integer"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "City not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "City": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Cairo"
                    },
                    "name_ar": {
                        "type": "string",
                        "example": "القاهرة"
                    },
                    "slug": {
                        "type": "string",
                        "example": "cairo"
                    },
                    "description": {
                        "type": "string",
                        "nullable": true
                    },
                    "description_ar": {
                        "type": "string",
                        "nullable": true
                    },
                    "image": {
                        "type": "string",
                        "nullable": true
                    },
                    "banner_image": {
                        "type": "string",
                        "nullable": true
                    },
                    "latitude": {
                        "type": "number",
                        "format": "float",
                        "example": 30.0444
                    },
                    "longitude": {
                        "type": "number",
                        "format": "float",
                        "example": 31.2357
                    },
                    "population": {
                        "type": "integer",
                        "nullable": true
                    },
                    "area_km2": {
                        "type": "number",
                        "format": "float",
                        "nullable": true
                    },
                    "is_capital": {
                        "type": "boolean",
                        "example": true
                    },
                    "is_active": {
                        "type": "boolean",
                        "example": true
                    },
                    "active_shops_count": {
                        "type": "integer",
                        "example": 150
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Enter your bearer token",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}